requirements.txt
python-dotenv==1.0.0
requests==2.31.0
polars==0.20.2
matplotlib==3.8.0
pyyaml==6.0.1 